Extract MdnsSource: zeroconf browser + cache accessors
Fourth PR in the device-state-monitor split arc. Mirrors the
legacy ``esphome/dashboard/status/mdns.py`` shape: an
:class:`MdnsSource` taking the monitor in ``__init__``, owning
the zeroconf responder, the ``AsyncServiceBrowser``, the
esphomelib service-state callback, and the cache accessors the
drawer's reachability subscription uses.

**``mdns.py``** — new ``MdnsSource`` class:

* ``start`` — was ``_start_mdns_browser``. Constructs
  ``AsyncEsphomeZeroconf``, ``DashboardImportDiscovery``, and
  the dispatch closure that routes esphomelib service-state
  events to its own ``_on_esphomelib_service_state_change``
  and HTTP events back to ``monitor._on_http_service_state_change``
  (HTTP / importable lives on the monitor for PR 4; PR 5 lifts
  it into ``ImportableDiscovery``).
* ``cancel_browser`` / ``close_zeroconf`` — split stop into
  two so the monitor's resolve-task drain still happens
  between them (the "cancel browser FIRST" invariant).
* ``_on_esphomelib_service_state_change`` — was the inner
  closure in ``_start_mdns_browser``; now a private method.
* ``_resolve_and_apply`` / ``_resolve_then`` /
  ``_apply_service_info`` — service-info resolve + apply
  helpers; ``_resolve_then`` is used by both the esphomelib
  path (here) and the HTTP path (still in controller.py
  via ``self._mdns._resolve_then``).
* ``refresh_mdns`` / ``_get_address_records`` /
  ``get_mdns_a_record_ttl_remaining`` / ``get_mdns_cache_info``
  / ``get_cached_addresses`` — the cache accessors.
* ``zeroconf`` property — exposes the responder for the
  monitor's own ``zeroconf`` property to delegate.

The ``_zeroconf`` / ``_mdns_browser`` fields move with their
class.

**``helpers.py``** — ``_HTTP_SERVICE_TYPE`` constant moves
here alongside ``_ESPHOME_SERVICE_TYPE`` so both controller.py
(``_seed_http_url_from_cache``, ``_on_http_service_state_change``)
and mdns.py (browser construction + dispatch) can import it.

**controller.py wiring**:

* ``__init__`` builds ``self._mdns = MdnsSource(self)``
  (alongside ``self._ping = PingSource(self)``).
* ``start`` → ``await self._mdns.start()``.
* ``stop`` → ``await self._mdns.cancel_browser()`` + drain
  + ``await self._mdns.close_zeroconf()`` preserving the
  cancel-browser-first invariant.
* Public mDNS surface (``refresh_mdns``,
  ``get_mdns_a_record_ttl_remaining``, ``get_mdns_cache_info``,
  ``get_cached_addresses``) becomes thin one-line delegates to
  ``self._mdns.*``; existing external callers
  (``controllers/devices/*``, tests reaching for the public
  API) keep working unchanged.
* ``zeroconf`` property delegates: ``return self._mdns.zeroconf``.
* HTTP / importable methods that still reference the responder
  (``probe_device``, ``_seed_http_url_from_cache``,
  ``_resolve_and_apply_http``) reach through
  ``self._mdns.zeroconf`` / ``self._mdns._resolve_then``;
  they'll move into ``ImportableDiscovery`` in PR 5.

**``shared.py``** — ``resolve_non_api_mdns_targets``'s
``monitor._zeroconf`` reference becomes ``monitor._mdns.zeroconf``.

**Test redirects** (mechanical):

* Patches against ``state_monitor_module.AsyncEsphomeZeroconf``
  / ``AsyncServiceBrowser`` / ``DashboardImportDiscovery`` /
  ``AddressResolver`` redirect to ``mdns_module``.
* ``AsyncServiceInfo`` patches redirect based on the test
  function name — ``test_dispatch_added_*`` go to
  ``mdns_module`` (esphomelib path); ``test_dispatch_http_*``
  keep ``state_monitor_module`` (HTTP path stays in
  controller.py).
* ``monitor._zeroconf`` / ``monitor._mdns_browser`` direct
  field access → ``monitor._mdns._zeroconf`` /
  ``monitor._mdns._mdns_browser``.
* ``monkeypatch.setattr(monitor, "_apply_service_info", …)``
  / ``"_resolve_and_apply"`` → ``monitor._mdns``.
* ``monitor._start_mdns_browser()`` → ``monitor._mdns.start()``.
* ``__new__``-built test fixtures now wire ``monitor._mdns =
  MdnsSource(monitor)`` after constructing ``MonitorState``.

controller.py: 1143 → 755 lines.
